[feature] Add policy filtering
Allow for filtered policy loading, so that the loaded policy is
a targeted subset of the policy in storage. This allows the policy
enforcement to more effectively scale for very large policies in
a multi-tenant environment by only loading the policies relevant
to the requested tenant.

To protect the full policy from accidental corruption, a filtered
policy cannot be saved back to storage.